+++ src/components/MyProfile/MyProfile.css
@@ -1,3 +1,12 @@
-h1 {
-  color: red;
+.my-profile-container {
+  display: grid;
+  grid-template-columns: 1fr 2fr 1fr; 
+  align-items: flex-start; 
+}
+
+.column {
+  padding: 20px;
+  border: 1px solid #ccc;
+  background-color: #fff;
+  box-shadow: 0 0 5px rgba(0, 0, 0, 0.1);
 }
